+++ src/test/java/step4/domain/lotto/LottoTest.java
@@ -102,7 +102,7 @@ class LottoTest {
         // when
         Lotto lotto = Lotto.of(lottoNumbers);
         Lotto anotherLotto = Lotto.of(anotherLottoNumbers);
-        int actual = lotto.getCountMatch(anotherLotto);
+        long actual = lotto.getCountMatch(anotherLotto);
 
         // then
         assertThat(actual).isEqualTo(1);
